requirements.txt
# Core Framework
fastapi==0.121.2
uvicorn[standard]==0.38.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20

# Google APIs
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0

# Video Download
yt-dlp==2024.12.23
httpx==0.28.1
aiofiles==24.1.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Security & Utils
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
cryptography==44.0.0
tenacity==9.0.0

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0
